requirements.txt
setuptools==41.0.0
tensorflow==1.15.0
protobuf==3.8.0
redis==3.0.1
jinja2
yappi
grpcio
grpcio-tools
Flask==1.0.2
requests
peewee==3.9.3
PyMySQL==0.9.3
Flask-SocketIO==3.0.2
cachetools
Werkzeug==0.15.3

